<< (int)pid; Some systems don't use ints for pids, but since the DCOP method uses int...
svn path=/trunk/kdelibs/; revision=46309

+++ kio/krun.cpp
@@ -744,7 +744,7 @@ void KRun::clientStarted(
 {
   QByteArray params;
   QDataStream stream(params, IO_WriteOnly);
-  stream << execName << iconName << resName << pid;
+  stream << execName << iconName << resName << (int)pid;
   kapp->dcopClient()->send(
     "kicker",
     "TaskbarApplet",
